update to new

+++ chrome/browser/ui/hung_plugin_tab_helper.cc
@@ -42,7 +42,7 @@ void KillPluginOnIOThread(int child_id) {
     if (data.id == child_id) {
       CrashDumpHungChildProcess(data.GetProcess().Handle());
       data.GetProcess().Terminate(content::RESULT_CODE_HUNG, false);
-      break;
+      return;
     }
     ++iter;
   }
@@ -52,7 +52,6 @@ void KillPluginOnIOThread(int child_id) {
 
 }  // namespace
 
-
 // HungPluginTabHelper::PluginState -------------------------------------------
 
 // Per-plugin state (since there could be more than one plugin hung).  The
@@ -61,72 +60,53 @@ void KillPluginOnIOThread(int child_id) {
 // not we're currently showing the infobar.
 struct HungPluginTabHelper::PluginState {
   // Initializes the plugin state to be a hung plugin.
-  PluginState(const base::FilePath& p, const base::string16& n);
-  ~PluginState();
+  PluginState(const base::FilePath& path, const base::string16& name);
+
+  // Since the scope of the timer manages our callback, this struct should
+  // not be copied.
+  PluginState(const PluginState&) = delete;
+  PluginState& operator=(const PluginState&) = delete;
+
+  ~PluginState() = default;
 
   base::FilePath path;
   base::string16 name;
 
   // Possibly-null if we're not showing an infobar right now.
-  infobars::InfoBar* infobar;
+  infobars::InfoBar* infobar = nullptr;
 
   // Time to delay before re-showing the infobar for a hung plugin. This is
   // increased each time the user cancels it.
-  base::TimeDelta next_reshow_delay;
+  base::TimeDelta next_reshow_delay = base::TimeDelta::FromSeconds(10);
 
   // Handles calling the helper when the infobar should be re-shown.
   base::OneShotTimer timer;
-
- private:
-  // Initial delay in seconds before re-showing the hung plugin message.
-  static const int kInitialReshowDelaySec;
-
-  // Since the scope of the timer manages our callback, this struct should
-  // not be copied.
-  DISALLOW_COPY_AND_ASSIGN(PluginState);
 };
 
-// static
-const int HungPluginTabHelper::PluginState::kInitialReshowDelaySec = 10;
-
-HungPluginTabHelper::PluginState::PluginState(const base::FilePath& p,
-                                              const base::string16& n)
-    : path(p),
-      name(n),
-      infobar(NULL),
-      next_reshow_delay(base::TimeDelta::FromSeconds(kInitialReshowDelaySec)) {}
-
-HungPluginTabHelper::PluginState::~PluginState() {
-}
-
+HungPluginTabHelper::PluginState::PluginState(const base::FilePath& path,
+                                              const base::string16& name)
+    : path(path), name(name) {}
 
 // HungPluginTabHelper --------------------------------------------------------
 
-HungPluginTabHelper::HungPluginTabHelper(content::WebContents* contents)
-    : content::WebContentsObserver(contents), infobar_observer_(this) {}
-
-HungPluginTabHelper::~HungPluginTabHelper() {
-}
+HungPluginTabHelper::~HungPluginTabHelper() = default;
 
 void HungPluginTabHelper::PluginCrashed(const base::FilePath& plugin_path,
                                         base::ProcessId plugin_pid) {
-  // TODO(brettw) ideally this would take the child process ID. When we do this
-  // for NaCl plugins, we'll want to know exactly which process it was since
-  // the path won't be useful.
-  InfoBarService* infobar_service =
-      InfoBarService::FromWebContents(web_contents());
-  if (!infobar_service)
-    return;
-
   // For now, just do a brute-force search to see if we have this plugin. Since
   // we'll normally have 0 or 1, this is fast.
-  for (auto i = hung_plugins_.begin(); i != hung_plugins_.end(); ++i) {
-    if (i->second->path == plugin_path) {
-      if (i->second->infobar)
+  const auto i = std::find_if(hung_plugins_.begin(), hung_plugins_.end(),
+                              [plugin_path](const auto& elem) {
+                                return elem.second->path == plugin_path;
+                              });
+  if (i != hung_plugins_.end()) {
+    if (i->second->infobar) {
+      InfoBarService* infobar_service =
+          InfoBarService::FromWebContents(web_contents());
+      if (infobar_service)
         infobar_service->RemoveInfoBar(i->second->infobar);
-      hung_plugins_.erase(i);
-      break;
     }
+    hung_plugins_.erase(i);
   }
 }
 
@@ -136,26 +116,26 @@ void HungPluginTabHelper::PluginHungStatusChanged(
     bool is_hung) {
   InfoBarService* infobar_service =
       InfoBarService::FromWebContents(web_contents());
-  if (!infobar_service)
-    return;
-  if (!infobar_observer_.IsObserving(infobar_service))
-    infobar_observer_.Add(infobar_service);
 
   auto found = hung_plugins_.find(plugin_child_id);
   if (found != hung_plugins_.end()) {
     if (!is_hung) {
       // Hung plugin became un-hung, close the infobar and delete our info.
-      if (found->second->infobar)
+      if (found->second->infobar && infobar_service)
         infobar_service->RemoveInfoBar(found->second->infobar);
       hung_plugins_.erase(found);
     }
     return;
   }
 
+  if (!infobar_service)
+    return;
+  if (!infobar_observer_.IsObserving(infobar_service))
+    infobar_observer_.Add(infobar_service);
+
   base::string16 plugin_name =
       content::PluginService::GetInstance()->GetPluginDisplayNameByPath(
           plugin_path);
-
   hung_plugins_[plugin_child_id] =
       std::make_unique<PluginState>(plugin_path, plugin_name);
   ShowBar(plugin_child_id, hung_plugins_[plugin_child_id].get());
@@ -163,22 +143,21 @@ void HungPluginTabHelper::PluginHungStatusChanged(
 
 void HungPluginTabHelper::OnInfoBarRemoved(infobars::InfoBar* infobar,
                                            bool animate) {
-  for (auto i = hung_plugins_.begin(); i != hung_plugins_.end(); ++i) {
+  const auto i = std::find_if(
+      hung_plugins_.begin(), hung_plugins_.end(),
+      [infobar](const auto& elem) { return elem.second->infobar == infobar; });
+  if (i != hung_plugins_.end()) {
     PluginState* state = i->second.get();
-    if (state->infobar == infobar) {
-      state->infobar = nullptr;
+    state->infobar = nullptr;
 
-      // Schedule the timer to re-show the infobar if the plugin continues to be
-      // hung.
-      state->timer.Start(FROM_HERE, state->next_reshow_delay,
-          base::Bind(&HungPluginTabHelper::OnReshowTimer,
-                     base::Unretained(this),
-                     i->first));
+    // Schedule the timer to re-show the infobar if the plugin continues to be
+    // hung.
+    state->timer.Start(FROM_HERE, state->next_reshow_delay,
+                       base::BindOnce(&HungPluginTabHelper::OnReshowTimer,
+                                      base::Unretained(this), i->first));
 
-      // Next time we do this, delay it twice as long to avoid being annoying.
-      state->next_reshow_delay *= 2;
-      return;
-    }
+    // Next time we do this, delay it twice as long to avoid being annoying.
+    state->next_reshow_delay *= 2;
   }
 }
 
@@ -188,13 +167,12 @@ void HungPluginTabHelper::OnManagerShuttingDown(
 }
 
 void HungPluginTabHelper::KillPlugin(int child_id) {
-  auto found = hung_plugins_.find(child_id);
-  DCHECK(found != hung_plugins_.end());
-
-  base::PostTaskWithTraits(FROM_HERE, {content::BrowserThread::IO},
-                           base::BindOnce(&KillPluginOnIOThread, child_id));
-  CloseBar(found->second.get());
+  base::PostTask(FROM_HERE, {content::BrowserThread::IO},
+                 base::BindOnce(&KillPluginOnIOThread, child_id));
 }
+
+HungPluginTabHelper::HungPluginTabHelper(content::WebContents* contents)
+    : content::WebContentsObserver(contents) {}
 
 void HungPluginTabHelper::OnReshowTimer(int child_id) {
   // The timer should have been cancelled if the record isn't in our map
@@ -214,15 +192,6 @@ void HungPluginTabHelper::ShowBar(int child_id, PluginState* state) {
   DCHECK(!state->infobar);
   state->infobar = HungPluginInfoBarDelegate::Create(infobar_service, this,
                                                      child_id, state->name);
-}
-
-void HungPluginTabHelper::CloseBar(PluginState* state) {
-  InfoBarService* infobar_service =
-      InfoBarService::FromWebContents(web_contents());
-  if (infobar_service && state->infobar) {
-    infobar_service->RemoveInfoBar(state->infobar);
-    state->infobar = NULL;
-  }
 }
 
 WEB_CONTENTS_USER_DATA_KEY_IMPL(HungPluginTabHelper)
