+++ testing/test_env.py
@@ -86,6 +86,9 @@ def get_sanitizer_env(cmd, asan, lsan, msan, tsan, cfi_diag):
   else:
     symbolization_options = []
 
+  # Leverage sanitizer to print stack trace on abort (e.g. assertion failure).
+  symbolization_options.append('handle_abort=1')
+
   if asan:
     asan_options = symbolization_options[:]
     if lsan:
@@ -186,8 +189,8 @@ def run_command_with_output(argv, stdoutfile, env=None, cwd=None):
   assert stdoutfile
   with io.open(stdoutfile, 'wb') as writer, \
       io.open(stdoutfile, 'rb', 1) as reader:
-    process = subprocess.Popen(argv, env=env, cwd=cwd, stdout=writer,
-        stderr=subprocess.STDOUT)
+    process = _popen(argv, env=env, cwd=cwd, stdout=writer,
+                     stderr=subprocess.STDOUT)
     forward_signals([process])
     while process.poll() is None:
       sys.stdout.write(reader.read())
@@ -211,9 +214,27 @@ def run_command(argv, env=None, cwd=None, log=True):
   """
   if log:
     print('Running %r in %r (env: %r)' % (argv, cwd, env))
-  process = subprocess.Popen(argv, env=env, cwd=cwd, stderr=subprocess.STDOUT)
+  process = _popen(argv, env=env, cwd=cwd, stderr=subprocess.STDOUT)
   forward_signals([process])
   return wait_with_signals(process)
+
+
+def run_command_output_to_handle(argv, file_handle, env=None, cwd=None):
+  """Run command and stream its stdout/stderr both to |file_handle|.
+
+  Also forward_signals to obey
+  https://chromium.googlesource.com/infra/luci/luci-py/+/master/appengine/swarming/doc/Bot.md#graceful-termination_aka-the-sigterm-and-sigkill-dance
+
+  Returns:
+    integer returncode of the subprocess.
+  """
+  print('Running %r in %r (env: %r)' % (argv, cwd, env))
+  process = _popen(
+      argv, env=env, cwd=cwd, stderr=file_handle, stdout=file_handle)
+  forward_signals([process])
+  exit_code = wait_with_signals(process)
+  print('Command returned exit code %d' % exit_code)
+  return exit_code
 
 
 def wait_with_signals(process):
@@ -250,6 +271,8 @@ def forward_signals(procs):
   assert all(isinstance(p, subprocess.Popen) for p in procs)
   def _sig_handler(sig, _):
     for p in procs:
+      if p.poll() is not None:
+        continue
       # SIGBREAK is defined only for win32.
       if sys.platform == 'win32' and sig == signal.SIGBREAK:
         p.send_signal(signal.CTRL_BREAK_EVENT)
@@ -259,6 +282,7 @@ def forward_signals(procs):
     signal.signal(signal.SIGBREAK, _sig_handler)
   else:
     signal.signal(signal.SIGTERM, _sig_handler)
+    signal.signal(signal.SIGINT, _sig_handler)
 
 
 def run_executable(cmd, env, stdoutfile=None):
@@ -311,10 +335,17 @@ def run_executable(cmd, env, stdoutfile=None):
   cmd[0] = cmd[0].replace('/', os.path.sep)
   cmd = fix_python_path(cmd)
 
+  # We also want to print the GTEST env vars that were set by the caller,
+  # because you need them to reproduce the task properly.
+  env_to_print = extra_env.copy()
+  for env_var_name in ('GTEST_SHARD_INDEX', 'GTEST_TOTAL_SHARDS'):
+      if env_var_name in env:
+          env_to_print[env_var_name] = env[env_var_name]
+
   print('Additional test environment:\n%s\n'
         'Command: %s\n' % (
         '\n'.join('    %s=%s' %
-            (k, v) for k, v in sorted(extra_env.iteritems())),
+            (k, v) for k, v in sorted(env_to_print.iteritems())),
         ' '.join(cmd)))
   sys.stdout.flush()
   env.update(extra_env or {})
@@ -325,9 +356,9 @@ def run_executable(cmd, env, stdoutfile=None):
     elif use_symbolization_script:
       # See above comment regarding offline symbolization.
       # Need to pipe to the symbolizer script.
-      p1 = subprocess.Popen(cmd, env=env, stdout=subprocess.PIPE,
-                            stderr=sys.stdout)
-      p2 = subprocess.Popen(
+      p1 = _popen(cmd, env=env, stdout=subprocess.PIPE,
+                  stderr=sys.stdout)
+      p2 = _popen(
           get_sanitizer_symbolize_command(executable_path=cmd[0]),
           env=env, stdin=p1.stdout)
       p1.stdout.close()  # Allow p1 to receive a SIGPIPE if p2 exits.
@@ -342,6 +373,14 @@ def run_executable(cmd, env, stdoutfile=None):
   except OSError:
     print >> sys.stderr, 'Failed to start %s' % cmd
     raise
+
+
+def _popen(*args, **kwargs):
+  assert 'creationflags' not in kwargs
+  if sys.platform == 'win32':
+    # Necessary for signal handling. See crbug.com/733612#c6.
+    kwargs['creationflags'] = subprocess.CREATE_NEW_PROCESS_GROUP
+  return subprocess.Popen(*args, **kwargs)
 
 
 def main():
